Add files via upload
This is a number game, in this the computer will generate a number the user has to guess that number. The user has 3 chances to guess the number. If the user guesses the number in 1st chance the user will have 100 score, if the user guesses the number in 2nd chance the user will have 50 score, if the user guesses the number in 3rd chance the user will have 25 score and. If the user did not guess the number with in 3 chances, the user loss the game.